requirements.txt
pandas
python-dotenv
loguru
requests